requirements.txt
python-six
textblob==0.15.3
